server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
supabase==1.2.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
scrapy==2.11.0
playwright==1.41.2
beautifulsoup4==4.12.2
transformers==4.36.2
torch==2.1.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
aiohttp==3.9.1
python-slugify==8.0.1